refactor: 로직 분리에 맞게 테스트 코드 리팩토링
