+++ wabi/src/test/kotlin/com/wap/wabi/common/TestConstants.kt
@@ -2,5 +2,6 @@ package com.wap.wabi.common
 
 object TestConstants {
     const val ADMIN_ID = 1L
+    const val ANOTHER_ADMIN_ID = 2L
     const val CLUB = "Test Club"
 }
